+++ plc4go/internal/plc4go/ads/readwrite/model/AdsMultiRequestItemRead.go
@@ -45,7 +45,7 @@ type IAdsMultiRequestItemRead interface {
 // Accessors for discriminator values.
 ///////////////////////////////////////////////////////////
 func (m *AdsMultiRequestItemRead) IndexGroup() uint32 {
-	return 61568
+	return uint32(61568)
 }
 
 func (m *AdsMultiRequestItemRead) InitializeParent(parent *AdsMultiRequestItem) {
