+++ plc4go/internal/plc4go/ads/readwrite/model/AdsMultiRequestItemReadWrite.go
@@ -46,7 +46,7 @@ type IAdsMultiRequestItemReadWrite interface {
 // Accessors for discriminator values.
 ///////////////////////////////////////////////////////////
 func (m *AdsMultiRequestItemReadWrite) IndexGroup() uint32 {
-	return 61570
+	return uint32(61570)
 }
 
 func (m *AdsMultiRequestItemReadWrite) InitializeParent(parent *AdsMultiRequestItem) {
