+++ plc4go/internal/plc4go/knxnetip/readwrite/model/KnxGroupAddress.go
@@ -99,11 +99,11 @@ func KnxGroupAddressParse(readBuffer utils.ReadBuffer, numLevels uint8) (*KnxGro
 	var _parent *KnxGroupAddress
 	var typeSwitchError error
 	switch {
-	case numLevels == 1: // KnxGroupAddressFreeLevel
+	case numLevels == uint8(1): // KnxGroupAddressFreeLevel
 		_parent, typeSwitchError = KnxGroupAddressFreeLevelParse(readBuffer, numLevels)
-	case numLevels == 2: // KnxGroupAddress2Level
+	case numLevels == uint8(2): // KnxGroupAddress2Level
 		_parent, typeSwitchError = KnxGroupAddress2LevelParse(readBuffer, numLevels)
-	case numLevels == 3: // KnxGroupAddress3Level
+	case numLevels == uint8(3): // KnxGroupAddress3Level
 		_parent, typeSwitchError = KnxGroupAddress3LevelParse(readBuffer, numLevels)
 	default:
 		// TODO: return actual type
