+++ plc4go/internal/plc4go/knxnetip/readwrite/model/KnxGroupAddress3Level.go
@@ -45,7 +45,7 @@ type IKnxGroupAddress3Level interface {
 // Accessors for discriminator values.
 ///////////////////////////////////////////////////////////
 func (m *KnxGroupAddress3Level) NumLevels() uint8 {
-	return 3
+	return uint8(3)
 }
 
 func (m *KnxGroupAddress3Level) InitializeParent(parent *KnxGroupAddress) {
