+++ plc4go/internal/plc4go/knxnetip/readwrite/model/KnxGroupAddressFreeLevel.go
@@ -43,7 +43,7 @@ type IKnxGroupAddressFreeLevel interface {
 // Accessors for discriminator values.
 ///////////////////////////////////////////////////////////
 func (m *KnxGroupAddressFreeLevel) NumLevels() uint8 {
-	return 1
+	return uint8(1)
 }
 
 func (m *KnxGroupAddressFreeLevel) InitializeParent(parent *KnxGroupAddress) {
